Tamil_News_Translator/requirements.txt

# Core web scraping and HTTP libraries
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# RSS feed parsing
feedparser>=6.0.10

# Google News URL decoding
googlenewsdecoder>=0.1.7

# Translation libraries
googletrans>=4.0.0
google-cloud-translate>=3.15.0

# Text processing
newspaper3k>=0.2.8

# Error handling and retries
tenacity>=8.2.0
urllib3>=2.0.0

# Data handling
pandas>=2.0.0

# Async support (optional)
aiohttp>=3.8.0
asyncio

# Development and testing
pytest>=7.0.0
black>=23.0.0

